feat: add fast version of strcapture
Implement fstrcapture, a faster replacement for strcapture when
using Perl-compatible regular expressions. This is a drop-in
replacement save for the change of the default argument for `perl`
from `FALSE` to `TRUE`.

Utilises code from both strcapture and nc::capture_first_vec so
add to bump the licence to account for this.