feat: changes to cc
`cc()` now allows you to split length one character vectors by whitespace.

BREAKING CHANGE:
`cc()` no longer allows the user to send generative code for the output
to the system clipboard with the `.clip` argument. This functionality
was unreliable on some systems so has been removed until a reliable
implementation can be provided. I think the issue is related to the
interaction of RStudio, clipr and wayland (it seemed to work as
expected when just using R from the terminal).